+++ Vector2D.h
@@ -5,7 +5,7 @@ class Vector2D
 {
 public:
    Vector2D() : x(0), y(0) {}
-   Vector2D(const int vx, const int vy) : x(vx), y(vy) {}
+   Vector2D(const int16_t vx, const int16_t vy) : x(vx), y(vy) {}
    Vector2D& operator+=(const Vector2D& v) { x += v.x; y += v.y; return *this; }
    Vector2D& operator-=(const Vector2D& v) { x -= v.x; y -= v.y; return *this; }
    Vector2D& operator*=(const Vector2D& v) { x *= v.x; y *= v.y; return *this; }
@@ -13,8 +13,8 @@ public:
    friend bool operator==(const Vector2D& left, const Vector2D& right);
 
 public:
-   int x;
-   int y;
+   int16_t x;
+   int16_t y;
 };
 
 inline std::ostream &operator<<(std::ostream &os, Vector2D const &m) {
